+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Vaults: custom pattern lock
+- Video: picture-in-picture
+- Video: handle skip next/previous media buttons
 
 ### Changed